+++ src/main/java/com/myprojects/HttpClient.java
@@ -1,7 +1,7 @@
 package com.myprojects;
 
-import com.myprojects.WeatherAPP_Singletones.LongNameSplitter;
-import com.myprojects.WeatherAPP_Singletones.PropertyGetter;
+import com.myprojects.Services.LongNameSplitter;
+import com.myprojects.Services.PropertyGetter;
 import org.apache.http.HttpEntity;
 import org.apache.http.client.methods.CloseableHttpResponse;
 import org.apache.http.client.methods.HttpGet;
